move navigator on top of canvas, rename split-view to design
